insert and replace event no longer has a return value
If we are returning the inserted or replaced events in aw-server it is
impossible to use transactions in the database as that makes the commit
of the events lazy so no ID is decided at the end of the request.

My suggestion is to remove the return body from the API completely as
that gives much better I/O performance for aw-server.