Heartbeats should not return anything else than success/failure
If we want to be able to return the returning heartbeat from the API we are
forced to commit the database on every heartbeat. This drains a lot of performance
and is disabled on aw-server-rust and on aw-server-python with the sqlite
datastore for this reason. Doing such small writes is especially damaging for
HDDs, but the wakeups for SSDs also take a lot of IO performance and battery.

We do not have any use-case for this feature today anyway, so I think we should drop it.